Move log functions in separated file
